git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@5100 ebf7c1c2-ba36-0410-9fe8-c592906822b4

+++ src/jd/nutils/JDImage.java
@@ -59,7 +59,7 @@ public class JDImage {
             return iconToImage(new JFileChooser().getIcon(file));
 
         } finally {
-            file.delete();
+            if(file!=null)            file.delete();
         }
 
     }
